requirements.txt
anthropic>=0.40.0
openai>=1.50.0
pyyaml>=6.0
python-frontmatter>=1.1.0
aiohttp>=3.10.0
websockets>=13.0
rich>=13.0
click>=8.1
schedule>=1.2.0
pytest>=8.0
pytest-asyncio>=0.24.0
